requirements.txt
djangorestframework
django-cors-headers
pillow
django-nested-inline
django-cleanup
django-ckeditor
django-filter
pytils

django-resized
drf-yasg



